requirements.txt
comtypes
absl-py==2.3.1
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.122.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-texttospeech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.8.0
mediapipe==0.10.14
ml_dtypes==0.5.3
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pefile==2023.2.7
pillow==10.2.0
proto-plus==1.26.1
protobuf==4.25.3
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.5
rsa==4.9.1
sbc==1.3
scipy==1.15.3
screen_brightness_control==0.24.3
sentencepiece==0.2.1
shiboken6==6.10.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.4
srt==3.5.3
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
WMI==1.5.1